vboot: Correcting argument type for VbExTpmSendReceive
Length argument for VbExTpmSendReceive has type uint32_t whereas it calls
function which uses the same variable with type size_t. Thus, correcting the
type of length argument from uint32_t to size_t.

BUG=None
BRANCH=None
TEST=Compiled successfully

+++ firmware/lib/tpm_lite/tlcl.c
@@ -63,7 +63,7 @@ static inline int TpmReturnCode(const uint8_t* buffer) {
 static uint32_t TlclSendReceiveNoRetry(const uint8_t* request,
                                        uint8_t* response, int max_length) {
 
-  uint32_t response_length = max_length;
+  size_t response_length = max_length;
   uint32_t result;
 
 #ifdef EXTRA_LOGGING
